pass 2660 tests

+++ general/srcs/flags_parser.c
@@ -1,18 +1,25 @@
-//
-// Created by Aletha Yellin on 22/11/2019.
-//
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   flags_parser.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ayellin <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/01/21 22:57:56 by ayellin           #+#    #+#             */
+/*   Updated: 2020/01/21 22:58:28 by ayellin          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "flags_parse.h"
 
-int is_flag(char c)
+int		is_flag(char c)
 {
 	return (ft_strchr(FT_PRINTF_FLAGS, c) != NULL);
 }
 
-//TODO: if flags are repeated?
-int parse_flags(const char *format, t_spec *spec)
+int		parse_flags(const char *format, t_spec *spec)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	while (is_flag(format[i]))
